+++ lib/src/ui/has_alignment.dart
@@ -1,0 +1,12 @@
+//Copyright (C) 2012 Sergey Akopkokhyants. All Rights Reserved.
+//Author: akserg
+
+part of dart_web_toolkit_ui;
+
+/**
+ * A widget that implements this interface has an intrinsic alignment which
+ * can be manipulated using these methods and constants.
+ */
+abstract class HasAlignment implements HasHorizontalAlignment, HasVerticalAlignment {
+  
+}